pyproject.toml
[project]
name = "vectra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.15",
    "google-genai>=1.24.0",
    "questionary>=2.1.0",
    "rich>=14.0.0",
]

[project.scripts]
vectra = "vectra.__main__:main"
